refactor(server): split HTTP transport into its own TU (F1)
otlp_server.cpp mixed the buffered-group-commit ingest engine with the
httplib HTTP transport in one 1074-line file. Move the transport (ctor,
request handlers, listener, Impl, wasm stubs) to otlp_server_http.cpp so
httplib.hpp is confined to one TU; the engine TU no longer includes it.

OtlpHttpError (thrown in the engine, caught in the handler) and
OtlpSignalBuffer (owned by the class, destroyed in the ctor/dtor which now
live in the transport TU) move to a shared otlp_server_internal.hpp so both
TUs see the complete types.

Pure relocation, no logic change (engine diff: 274 deletions, 1 insertion).
The engine TU stays one cohesive unit (~770 LoC): splitting buffer-vs-seal
would cut across tightly shared state (signal_buffers, admission counters,
mutexes) for no real decoupling.
