+++ CMakeLists.txt
@@ -25,6 +25,7 @@ set(EXTENSION_SOURCES
     src/otlp_arrow.cpp
     src/otlp_uri.cpp
     src/otlp_server.cpp
+    src/otlp_server_http.cpp
     src/otlp_storage.cpp
     src/otlp_start_stop.cpp)
