+++ src/otlp_server.cpp
@@ -13,12 +13,9 @@
 
 #include "otlp_arrow.hpp"
 #include "otlp_log.hpp"
+#include "otlp_server_internal.hpp"
 
 #include <chrono>
-
-#ifndef __EMSCRIPTEN__
-#include "httplib.hpp"
-#endif
 
 namespace duckdb {
 
@@ -82,47 +79,6 @@ static void RunSQL(Connection &connection, const string &sql) {
 	}
 }
 
-class OtlpHttpError : public std::exception {
-public:
-	OtlpHttpError(int status_p, string message_p) : status(status_p), message(std::move(message_p)) {
-	}
-
-	const char *what() const noexcept override {
-		return message.c_str();
-	}
-
-	int status;
-	string message;
-};
-
-static string JsonEscape(const string &input) {
-	string result;
-	result.reserve(input.size() + 8);
-	for (auto c : input) {
-		switch (c) {
-		case '\\':
-			result += "\\\\";
-			break;
-		case '"':
-			result += "\\\"";
-			break;
-		case '\n':
-			result += "\\n";
-			break;
-		case '\r':
-			result += "\\r";
-			break;
-		case '\t':
-			result += "\\t";
-			break;
-		default:
-			result += c;
-			break;
-		}
-	}
-	return result;
-}
-
 static OtlpInputFormat FormatFromContentType(const string &content_type) {
 	auto value = content_type;
 	auto semicolon = value.find(';');
@@ -163,26 +119,6 @@ static bool IsUnsupportedEncoding(const string &content_encoding) {
 }
 
 } // namespace
-
-//! In-memory buffer for one OTLP signal table. The metadata fields are immutable
-//! after construction (so workers may read `types` lock-free); mutable rows live
-//! behind this signal's own mutex so independent signals don't contend on append.
-struct OtlpSignalBuffer {
-	const OtlpSignalType signal_type;
-	const string table_name;
-	const vector<LogicalType> types;
-	std::mutex mutex;
-	unique_ptr<ColumnDataCollection> collection;
-	idx_t buffered_rows = 0;
-	bool have_unsealed = false;
-	std::chrono::steady_clock::time_point first_unsealed;
-
-	OtlpSignalBuffer(OtlpSignalType signal_type_p, string table_name_p, vector<LogicalType> types_p,
-	                 unique_ptr<ColumnDataCollection> collection_p)
-	    : signal_type(signal_type_p), table_name(std::move(table_name_p)), types(std::move(types_p)),
-	      collection(std::move(collection_p)) {
-	}
-};
 
 void OtlpServer::GetSignalColumns(OtlpSignalType signal_type, vector<LogicalType> &types, vector<string> &names) {
 	ArrowSchema arrow_schema;
@@ -829,214 +765,5 @@ void OtlpServer::ShutdownIngest() {
 	}
 	writer_con.reset();
 }
-
-#ifndef __EMSCRIPTEN__
-
-class OtlpServer::Impl {
-public:
-	unique_ptr<duckdb_httplib::Server> server;
-};
-
-static void SetJson(duckdb_httplib::Response &res, int status, const string &json) {
-	res.status = status;
-	res.set_content(json, "application/json");
-}
-
-static void SetError(duckdb_httplib::Response &res, int status, const string &reason, const string &message) {
-	SetJson(res, status, "{\"error\":\"" + JsonEscape(reason) + "\",\"message\":\"" + JsonEscape(message) + "\"}");
-}
-
-OtlpServer::OtlpServer(ClientContext &context, const OtlpUri &uri_p, const OtlpServerConfig &config_p)
-    : db_ptr(context.db), uri(uri_p), config(config_p), impl(make_uniq<Impl>()) {
-	ValidateToken(config.token);
-	auto db = db_ptr.lock();
-	if (!db) {
-		throw InternalException("Database was closed");
-	}
-	EnsureTargetTables();
-	InitBuffers();
-	// One dedicated writer connection; the single sealer thread is the only thing
-	// that ever commits, so concurrent DuckLake writes never conflict.
-	writer_con = make_uniq<Connection>(*db);
-	writer_con->context->config.enable_progress_bar = false;
-	StartSealer();
-
-	impl->server = make_uniq<duckdb_httplib::Server>();
-	// Wide worker pool (mirrors duckdb-quack): each keep-alive connection holds a
-	// worker thread for its lifetime, so we need enough threads to serve many
-	// concurrent exporters at once. Workers only parse/convert and buffer rows in
-	// memory (lock-free apart from a brief buffer append); a single background sealer
-	// group-commits the buffer, so concurrent exporters never contend on the writer.
-	impl->server->new_task_queue = [] {
-		return new duckdb_httplib::ThreadPool(128);
-	};
-	impl->server->set_keep_alive_max_count(128);
-	impl->server->set_keep_alive_timeout(10);
-	impl->server->set_tcp_nodelay(true);
-	impl->server->set_payload_max_length(static_cast<size_t>(config.max_body_bytes));
-
-	impl->server->Get("/healthz", [](const duckdb_httplib::Request &, duckdb_httplib::Response &res) {
-		SetJson(res, 200, "{\"status\":\"ok\"}");
-	});
-
-	auto post_handler = [&](OtlpRequestKind kind) {
-		return [this, kind](const duckdb_httplib::Request &req, duckdb_httplib::Response &res) {
-			// RAII so active_requests is decremented on every exit path, including a
-			// throw that escapes the catch chain below (e.g. allocation failure).
-			struct RequestGuard {
-				std::atomic<idx_t> &counter;
-				explicit RequestGuard(std::atomic<idx_t> &counter_p) : counter(counter_p) {
-					counter++;
-				}
-				~RequestGuard() {
-					counter--;
-				}
-			} request_guard {active_requests};
-			total_requests++;
-			try {
-				if (!CheckAuth(req.get_header_value("Authorization"), req.get_header_value("x-api-key"))) {
-					SetError(res, 401, "unauthorized", "missing or invalid Authorization bearer token or x-api-key");
-				} else {
-					auto result = Ingest(kind, req.get_header_value("Content-Type"),
-					                     req.get_header_value("Content-Encoding"), req.body);
-					// 202 Accepted: rows are parsed + buffered in memory, not yet durable.
-					// They commit at the next seal (otlp_flush / otlp_stop force one).
-					auto response =
-					    StringUtil::Format("{\"status\":\"buffered\",\"rows\":%llu,\"batches\":%llu",
-					                       static_cast<uint64_t>(result.rows), static_cast<uint64_t>(result.batches));
-					if (kind == OtlpRequestKind::METRICS || result.HasSkipped()) {
-						response += StringUtil::Format(
-						    ",\"skipped\":{\"summaries\":%llu,\"nan_values\":%llu,\"infinity_values\":%llu,"
-						    "\"missing_values\":%llu}",
-						    static_cast<uint64_t>(result.skipped_summaries),
-						    static_cast<uint64_t>(result.skipped_nan_values),
-						    static_cast<uint64_t>(result.skipped_infinity_values),
-						    static_cast<uint64_t>(result.skipped_missing_values));
-					}
-					response += "}";
-					SetJson(res, 202, response);
-				}
-			} catch (OtlpHttpError &ex) {
-				SetError(res, ex.status, "request_failed", ex.what());
-			} catch (InvalidInputException &ex) {
-				SetError(res, 400, "bad_request", ex.what());
-			} catch (IOException &ex) {
-				// 500-class failures are otherwise only visible in the client's
-				// response body; log them so the operator has a server-side trace.
-				LogServerEvent(StringUtil::Format("ingest I/O error: %s", ex.what()));
-				SetError(res, 500, "internal_error", ex.what());
-			} catch (std::exception &ex) {
-				LogServerEvent(StringUtil::Format("ingest internal error: %s", ex.what()));
-				SetError(res, 500, "internal_error", ex.what());
-			}
-		};
-	};
-
-	impl->server->Post("/v1/logs", post_handler(OtlpRequestKind::LOGS));
-	impl->server->Post("/v1/traces", post_handler(OtlpRequestKind::TRACES));
-	impl->server->Post("/v1/metrics", post_handler(OtlpRequestKind::METRICS));
-
-	// The sealer thread is already running. If bind fails (e.g. EADDRINUSE) this object
-	// is not fully constructed, so ~OtlpServer() won't run — stop the sealer here so a
-	// joinable std::thread is never destroyed (which would std::terminate the process).
-	try {
-		if (!impl->server->is_valid()) {
-			throw IOException("Failed to instantiate OTLP HTTP server at %s / %s", uri.Uri(), uri.Http());
-		}
-		// Bind synchronously here so that bind() failures (e.g. EADDRINUSE) propagate
-		// to the caller of otlp_serve() rather than being lost on the listener thread.
-		if (!impl->server->bind_to_port(uri.Host(), uri.Port())) {
-			throw IOException(
-			    "Failed to bind OTLP HTTP server to %s (address in use, permission denied, or invalid host/"
-			    "port)",
-			    uri.Http());
-		}
-	} catch (...) {
-		ShutdownIngest();
-		throw;
-	}
-	is_running.store(true);
-	listen_thread = std::thread(ListenThread, this);
-}
-
-void OtlpServer::StopAccepting() {
-	// Closes the listening socket only. Idempotent. Safe to call from a
-	// request-handler thread — does not wait on httplib's task queue.
-	if (is_running.exchange(false)) {
-		impl->server->stop();
-	}
-}
-
-void OtlpServer::Close() {
-	// Stops accepting new connections AND joins the listener threads (NOT the
-	// httplib worker pool). Must not be called from a worker thread — the
-	// listener's exit path inside httplib joins all workers, so a worker
-	// joining the listener would deadlock through that chain.
-	StopAccepting();
-	if (listen_thread.joinable()) {
-		listen_thread.join();
-	}
-	// Workers are now joined: stop the sealer and drain the remaining buffer before
-	// the writer connection / database go away. Safe here (controlling thread, not a
-	// worker); idempotent with the ~OtlpServer() safety-net call.
-	ShutdownIngest();
-}
-
-OtlpServer::~OtlpServer() {
-	try {
-		Close();
-	} catch (std::exception &) {
-	}
-}
-
-void OtlpServer::ListenThread(OtlpServer *server) {
-	// The socket is already bound (synchronously, in the constructor); this only
-	// runs the accept loop. Catch everything so the listener thread never lets an
-	// exception escape — that would call std::terminate and abort the host process.
-	try {
-		server->impl->server->listen_after_bind();
-	} catch (std::exception &ex) {
-		server->is_running.store(false);
-		server->listener_failed.store(true);
-		{
-			std::lock_guard<std::mutex> lock(server->error_mutex);
-			server->last_error = ex.what();
-		}
-		server->LogServerEvent(
-		    StringUtil::Format("OTLP listener for %s stopped: %s", server->ListenUri().Uri(), ex.what()));
-	} catch (...) {
-		server->is_running.store(false);
-		server->listener_failed.store(true);
-		{
-			std::lock_guard<std::mutex> lock(server->error_mutex);
-			server->last_error = "unknown error in listen loop";
-		}
-		server->LogServerEvent(
-		    StringUtil::Format("OTLP listener for %s stopped: unknown error", server->ListenUri().Uri()));
-	}
-}
-
-#else
-
-class OtlpServer::Impl {};
-
-OtlpServer::OtlpServer(ClientContext &context, const OtlpUri &uri_p, const OtlpServerConfig &config_p)
-    : db_ptr(context.db), uri(uri_p), config(config_p), impl(make_uniq<Impl>()) {
-	throw NotImplementedException("otlp_serve is not implemented for the wasm platform");
-}
-
-void OtlpServer::StopAccepting() {
-}
-
-void OtlpServer::Close() {
-}
-
-OtlpServer::~OtlpServer() {
-}
-
-void OtlpServer::ListenThread(OtlpServer *server) {
-}
-
-#endif
 
 } // namespace duckdb
